Changed the name of the driver to Xilinx DMA Bridge.
The driver's name has been changed from AXI DMA to Xilinx DMA Bridge.
The name AXI DMA caused a fair bit of confusion for two reasons. One,
people believed that it was low-level driver for Xilinx's DMA hardware
IPs, instead of being a bridge. Second, the name doesn't reflect that
AXI VDMA is also supported by the driver. The name bridge was chosen
because it more accurately reflects the driver's purpose as a bridge
between userspace and the Xilinx DMA drivers.